Configure the database
Now that we're using [Postgres] as our database we can add the database
configuration to the project. The [tox] configuration is being updated
to accept pass the database-related environment variables through to
each environment, using default values for variables that don't exist.
The [GitHub Actions][actions] configuration is being updated both to run
[Postgres] and to set the environment variables that [tox] needs.

[actions]: https://docs.github.com/en/actions
[postgres]: https://www.postgresql.org
[tox]: https://tox.readthedocs.io