Set the Python version to 3.9
While tox and GitHub Actions are already configured to use 3.9, there
are other ways to interact with this code base. Dependabot has submitted
a pull request (#151) that keeps failing CI because it is removing one
of pytest's dependencies. This dependency is a backport of a feature
that was added in Python 3.11. That means Dependabot is running under
3.11, not 3.9 like it should.

According to [this response][issue], the Python version used by
Dependabot can be controlled by adding a `.python-version` file. This
will add that file with the most recent version of 3.9.

[issue]: https://github.com/dependabot/dependabot-core/issues/1455#issuecomment-545389290